+++ examples/singleton.js
@@ -1,0 +1,5 @@
+const {singleton: config} = require("../lib/index");
+
+module.exports = () => {
+    return config.envVars.bool;
+};