+++ examples/validator.js
@@ -1,0 +1,10 @@
+const ukPostcode = {
+    name: "ukPostcode",
+    fn: (value) => {
+        return /([Gg][Ii][Rr] 0[Aa]{2})|((([A-Za-z][0-9]{1,2})|(([A-Za-z][A-Ha-hJ-Yj-y][0-9]{1,2})|(([A-Za-z][0-9][A-Za-z])|([A-Za-z][A-Ha-hJ-Yj-y][0-9][A-Za-z]?))))\s?[0-9][A-Za-z]{2})/.test(value);
+    }
+};
+
+module.exports = {
+    ukPostcode
+};